Rearrange types
