+++ src/layout/src/lib.rs
@@ -2,5 +2,5 @@ mod column;
 pub mod content;
 mod wrap;
 
-pub use column::{Column, ColumnPair, WithLineNumber, WithLineNumberFiller};
+pub use column::{Column, ColumnPair, Lineable, WithLineNumber, WithLineNumberFiller};
 pub use content::{Highlight, Highlighted, InlineParts, StyledContent};
